Require `GraphMakie` explicitly to plot
Warn user if `Makie` installed but not `GraphMakie`
